servidor_web/requirements.txt
flask
psycopg2-binary
jinja2
gunicorn